Componentize coins page banner.
