+++ app/javascript/App/bundles/CoinIndex/index.js
@@ -1,52 +1,10 @@
 import React, { Component, Fragment } from 'react'
 import { Table } from 'antd'
 import ColumnNames from './ColumnNames'
-import Button from '@material-ui/core/Button'
-import Paper from '@material-ui/core/Paper'
-import Typography from '@material-ui/core/Typography'
-import Grid from '@material-ui/core/Grid'
 import SearchCoins from '~/bundles/common/components/SearchCoins'
+import Banner from '~/bundles/common/components/Banner'
 import API from '../../bundles/common/utils/localAPI'
 import * as _ from 'lodash'
-import { withStyles, createStyles } from '@material-ui/core/styles'
-
-const styles = (theme) =>
-  createStyles({
-    mainFeaturedPost: {
-      backgroundColor: '#253640',
-      color: '#d7d7d7',
-      marginBottom: theme.spacing.unit * 4,
-      boxShadow: 'none',
-    },
-    mainFeaturedPostContent: {
-      padding: `${theme.spacing.unit * 4}px`,
-      [theme.breakpoints.up('sm')]: {
-        paddingLeft: 0,
-        paddingRight: 0,
-      },
-    },
-    header: {
-      marginBottom: '0.5em !important',
-      fontSize: '1.5rem',
-      textAlign: 'center',
-    },
-    subheader: {
-      fontSize: '0.8rem',
-      textAlign: 'center',
-      [theme.breakpoints.up('sm')]: {
-        fontSize: '1rem',
-      },
-    },
-    cta: {
-      backgroundColor: '#23adf0',
-      color: '#fff',
-      borderRadius: '2px',
-      '&:hover': {
-        backgroundColor: '#23adf0',
-        boxShadow: '0 2px 10px rgba(14, 151, 255, 0.4)',
-      },
-    },
-  })
 
 class CoinIndex extends Component {
   constructor(props) {
@@ -99,45 +57,9 @@ class CoinIndex extends Component {
     })
   }
   render() {
-    const { classes } = this.props
-
     return (
       <Fragment>
-        <Paper className={classes.mainFeaturedPost} square={true}>
-          <Grid
-            container
-            direction="column"
-            justify="center"
-            alignItems="center"
-            className={classes.mainFeaturedPostContent}
-          >
-            <Grid item md>
-              <Typography className={classes.header} color="inherit">
-                Financial intelligence for cryptocurrency
-              </Typography>
-            </Grid>
-            <Grid item md>
-              <Typography
-                className={classes.subheader}
-                color="inherit"
-                paragraph
-              >
-                Uncover buy and sell opportunities through data science backed
-                blockchain analytics.
-              </Typography>
-            </Grid>
-            <Grid item md>
-              <Button
-                className={classes.cta}
-                onClick={() => {
-                  window.location.href = `/signals`
-                }}
-              >
-                Learn more
-              </Button>
-            </Grid>
-          </Grid>
-        </Paper>
+        <Banner />
 
         <div className="flex">
           <h1 className="pt3 pl3">Coins</h1>
@@ -169,4 +91,4 @@ class CoinIndex extends Component {
   }
 }
 
-export default withStyles(styles)(CoinIndex)
+export default CoinIndex
